Address testing review comments.
